Creating a server for an auto delearship
Hello,

I am in need of some assistance.  As a side project I have been developing a website for a family members auto dealership.  So far I have created the website and I have also created a secondary site, which is accessible through a log in which houses a form of information about a certain vehicle.  With no back end experience I would like some advice on how I can get the form data to add to the inventory section of the main website.  I will attach the code for each side on separate branches.  Thanks for any advice!