Inventory Form
This form would be used to add images of vehicles, an apply information on said vehicle.  The goal is to get said information migrated to the main page i.e. the dealership employees would be able to log in on the main page, which brings them here, fill in the form, the data is then displayed on the inventory section of the main page for customers to view the various cars in inventory at that time.  Additionally, the employees would need to be able to remove entries as they are sold.